Fix wizard import/export
